crear relacion carrito producto (tiene un error)

+++ src/main/java/com/santiago/carrito_compras/Controllers/ProductController.java
@@ -1,7 +1,6 @@
 package com.santiago.carrito_compras.Controllers;
 
-import com.santiago.carrito_compras.Dto.ProductDao;
-import com.santiago.carrito_compras.Dto.UserDto;
+import com.santiago.carrito_compras.Dto.ProductDto;
 import com.santiago.carrito_compras.Entities.Product;
 import com.santiago.carrito_compras.Services.ProductService;
 import org.springframework.http.HttpStatus;
@@ -20,27 +19,37 @@ public class ProductController {
     }
 
     @GetMapping("/products")
-    public List<Product> searchAllProducts() {
+    public List<ProductDto> searchAllProducts() {
         return service.searchAllProducts();
     }
 
     @PostMapping("/products/{id}")
-    public Product searchProduct(@PathVariable long id) {
-        return service.findProductById(id);
+    public ResponseEntity<?> searchProduct(@PathVariable long id) {
+        if (!service.existById(id)){
+            return new ResponseEntity<>("Product not found by id " + id, HttpStatus.NOT_FOUND);
+        }
+        return new ResponseEntity<>(service.findProductInfoById(id), HttpStatus.OK);
     }
 
     @PostMapping ("/products/update/{id}")
-    public ResponseEntity<Product> updateProduct(@PathVariable Long id, @RequestBody ProductDao updatedProduct) {
+    public ResponseEntity<?> updateProduct(@PathVariable Long id, @RequestBody ProductDto updatedProduct) {
+        if (!service.existById(id)){
+            return new ResponseEntity<>("Product not found by id " + id, HttpStatus.NOT_FOUND);
+        }
         Product product = service.updateProductById(updatedProduct, id);
-        return ResponseEntity.ok(product);
+        return new ResponseEntity<>(product, HttpStatus.OK);
     }
     @PostMapping("/products/delete/{id}")
-    public void deleteProduct(@PathVariable long id){
+    public ResponseEntity<?> deleteProduct(@PathVariable long id){
+        if (!service.existById(id)){
+            return new ResponseEntity<>("Product not found by id " + id, HttpStatus.NOT_FOUND);
+        }
         service.deleteProductById(id);
+        return new ResponseEntity<>("Product deleted", HttpStatus.OK);
     }
 
     @PostMapping("/products/create")
-    public Product createProduct(@RequestBody ProductDao productDao){
-        return service.createProduct(productDao);
+    public ResponseEntity<?> createProduct(@RequestBody ProductDto productDao){
+        return new ResponseEntity<>(service.createProduct(productDao), HttpStatus.CREATED);
     }
 }
